Modify members type

+++ geometry/src/geometry/Point.java
@@ -11,9 +11,9 @@ package geometry;
  * @author BM
  */
 public class Point {
-    private int x, y, z;
+    private double x, y, z;
 
-    public Point(int x, int y, int z) {
+    public Point(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
